requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
chardet==3.0.4
Django==3.2.4
djangorestframework==3.12.4
PyMySQL==1.0.2
pyparsing==3.0.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.22.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.25.11
twitter-api-client==0.9.7